training-service/requirements.txt
fastapi
uvicorn
tensorflow
pandas
numpy
python-dotenv